Moved envoy related code under new common/envoy package
